Connect progress dialog to app detail dialog
